backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
email-validator==2.0.0
Pillow==10.0.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-magic==0.4.27
bcrypt==4.1.1
redis==5.0.1
celery==5.3.4
requests==2.31.0Flask-Limiter==3.5.0
